+++ src/test/resources/ru/vyarus/gradle/plugin/quality/java/sample/cpd/OtherStruct2.java
@@ -8,6 +8,11 @@ public class OtherStruct2 {
         Math.sqrt(12);
         Math.sqrt(12);
         Math.sqrt(12);
+        Math.sqrt(12);
+        Math.sqrt(12);
+        Math.sqrt(12);
+        Math.sqrt(12);
+        Math.sqrt(12);
     }
 
     public void differentMethod2() {